Inline integration test suite tagging
