--- integration_test/scripts/tag-test-suite.js
@@ -1,12 +1,0 @@
-function tagTestSuite(registerSuite, defaultTag) {
-  return (spec) => {
-    const describe = spec.describe.bind(spec);
-
-    spec.describe = (label, defineTests) =>
-      describe(label, defineTests, defaultTag);
-
-    return registerSuite(spec);
-  };
-}
-
-module.exports = { tagTestSuite };
